+++ portal-impl/src/com/liferay/portlet/wiki/service/impl/WikiPageLocalServiceImpl.java
@@ -19,6 +19,7 @@ import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.repository.model.FileEntry;
+import com.liferay.portal.kernel.repository.model.Folder;
 import com.liferay.portal.kernel.sanitizer.SanitizerUtil;
 import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.IndexerRegistryUtil;
@@ -54,6 +55,7 @@ import com.liferay.portlet.asset.NoSuchEntryException;
 import com.liferay.portlet.asset.model.AssetEntry;
 import com.liferay.portlet.asset.model.AssetLink;
 import com.liferay.portlet.asset.model.AssetLinkConstants;
+import com.liferay.portlet.documentlibrary.model.DLFolderConstants;
 import com.liferay.portlet.social.model.SocialActivity;
 import com.liferay.portlet.social.model.SocialActivityConstants;
 import com.liferay.portlet.trash.model.TrashEntry;
@@ -226,10 +228,12 @@ public class WikiPageLocalServiceImpl extends WikiPageLocalServiceBaseImpl {
 
 		WikiPage page = getPage(nodeId, title);
 
+		Folder folder = page.addAttachmentsFolder();
+
 		FileEntry fileEntry = PortletFileRepositoryUtil.addPortletFileEntry(
 			page.getGroupId(), userId, WikiPage.class.getName(),
-			page.getResourcePrimKey(), PortletKeys.WIKI,
-			page.getAttachmentsFolderId(), file, fileName, mimeType);
+			page.getResourcePrimKey(), PortletKeys.WIKI, folder.getFolderId(),
+			file, fileName, mimeType);
 
 		if (userId == 0) {
 			userId = page.getUserId();
@@ -254,10 +258,12 @@ public class WikiPageLocalServiceImpl extends WikiPageLocalServiceBaseImpl {
 
 		WikiPage page = getPage(nodeId, title);
 
+		Folder folder = page.addAttachmentsFolder();
+
 		FileEntry fileEntry = PortletFileRepositoryUtil.addPortletFileEntry(
 			page.getGroupId(), userId, WikiPage.class.getName(),
-			page.getResourcePrimKey(), PortletKeys.WIKI,
-			page.getAttachmentsFolderId(), inputStream, fileName, mimeType);
+			page.getResourcePrimKey(), PortletKeys.WIKI, folder.getFolderId(),
+			inputStream, fileName, mimeType);
 
 		if (userId == 0) {
 			userId = page.getUserId();
@@ -468,7 +474,11 @@ public class WikiPageLocalServiceImpl extends WikiPageLocalServiceBaseImpl {
 
 		// Attachments
 
-		PortletFileRepositoryUtil.deleteFolder(page.getAttachmentsFolderId());
+		long folderId = page.getAttachmentsFolderId();
+
+		if (folderId != DLFolderConstants.DEFAULT_PARENT_FOLDER_ID) {
+			PortletFileRepositoryUtil.deleteFolder(folderId);
+		}
 
 		// Subscriptions
 
@@ -539,8 +549,14 @@ public class WikiPageLocalServiceImpl extends WikiPageLocalServiceBaseImpl {
 
 		WikiPage page = getPage(nodeId, title);
 
+		long folderId = page.getAttachmentsFolderId();
+
+		if (folderId == DLFolderConstants.DEFAULT_PARENT_FOLDER_ID) {
+			return;
+		}
+
 		FileEntry fileEntry = PortletFileRepositoryUtil.getPortletFileEntry(
-			page.getGroupId(), page.getAttachmentsFolderId(), fileName);
+			page.getGroupId(), folderId, fileName);
 
 		deletePageAttachment(fileEntry.getFileEntryId());
 	}
@@ -550,8 +566,14 @@ public class WikiPageLocalServiceImpl extends WikiPageLocalServiceBaseImpl {
 
 		WikiPage page = getPage(nodeId, title);
 
+		long folderId = page.getAttachmentsFolderId();
+
+		if (folderId == DLFolderConstants.DEFAULT_PARENT_FOLDER_ID) {
+			return;
+		}
+
 		PortletFileRepositoryUtil.deletePortletFileEntries(
-			page.getGroupId(), page.getAttachmentsFolderId());
+			page.getGroupId(), folderId);
 	}
 
 	public void deletePages(long nodeId)
@@ -584,9 +606,14 @@ public class WikiPageLocalServiceImpl extends WikiPageLocalServiceBaseImpl {
 
 		WikiPage page = getPage(nodeId, title);
 
+		long folderId = page.getAttachmentsFolderId();
+
+		if (folderId == DLFolderConstants.DEFAULT_PARENT_FOLDER_ID) {
+			return;
+		}
+
 		PortletFileRepositoryUtil.deletePortletFileEntries(
-			page.getGroupId(), page.getAttachmentsFolderId(),
-			WorkflowConstants.STATUS_IN_TRASH);
+			page.getGroupId(), folderId, WorkflowConstants.STATUS_IN_TRASH);
 	}
 
 	public void discardDraft(long nodeId, String title, double version)
